Add files via upload
Keeping the description Short:

The dataset is used here is train.csv for toxic comment classification in the webpage of the comment section or anywhere. This Model can be applied anywhere to detect the toxicity of the comment. It will return you in a much-detailed way the level of toxicity of each and every word.